Rework from multiple binaries to a single binary using cli arguments.
- rename `generate_json.go` to `json.go`
- move `main()` from `generate_json.go` to `generate.go`
- update README
