+++ README.md
@@ -13,13 +13,24 @@ As an experiment the generation process is built using [Go](https://golang.org).
 
 The process of building the templates consists of two steps; generating JSON and generating HTML. The JSON files are committed to Git in the folder `/data`. This step should only be necessary when new content is made available.
 
+Before running any of the generators the application needs to be compiled. In order to prevent the need to install Go locally a Docker container is available:
+
+```bash
+docker-compose run go-cli go build -i -o generate
+```
+
+**NB** Due to the usage of a [CGO](https://golang.org/cmd/cgo/) module for connecting to Sqlite compilation can take about 30 seconds.
+
 ### Generating JSON sources
-First build the veekun/pokedex database following the [instructions](https://github.com/veekun/pokedex/wiki/Getting-Data). The database should be built to Sqlite. Make sure the resulting Sqlite data is copied to the working directory of this repository. The Sqlite database is dumped to `pokedex/data/pokedex.sqlite` by default.
+First build the veekun/pokedex database following the [instructions](https://github.com/veekun/pokedex/wiki/Getting-Data). The database should be built to Sqlite. Make sure the resulting Sqlite data is copied to the working directory of this repository. The Sqlite database is dumped to `pokedex/data/pokedex.sqlite` by default by Veekun.
 
-Next steps are preparation, compilation and generation:
-- `docker-compose run go-cli go build generate_json.go` - This command will compile the generation script.
-- `docker-compose run go-cli ./generate_json` - This command will run the compiled binary and will do the actual work.
+To (re)generate the json files run:
 
-After running these three commands a batch of JSON files should be created in `/data`.
+```bash
+docker-compose run go-cli ./generate json
+```
+
+After completion a batch of JSON files is written to `/data`.
 
 ### Generating HTML
+[WIP]
